seperated constatns and function from sdf client
